% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_dropbox_token}
\alias{get_dropbox_token}
\title{Get Dropbox token}
\usage{
get_dropbox_token()
}
\value{
character(1)
}
\description{
Get Dropbox token
}
\examples{
\dontrun{
  get_dropbox_token()
}
}
