% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{set_app_users}
\alias{set_app_users}
\title{Set app users}
\usage{
set_app_users(users, hash_passwords = TRUE, renviron = ".Renviron")
}
\arguments{
\item{users}{Data frame with columns user and password.}

\item{hash_passwords}{Logical.}

\item{renviron}{Path to .Renviron. Empty string disables file update.}
}
\value{
Character scalar.
}
\description{
Store users as JSON in APP_USERS_JSON.
Passwords can optionally be hashed using sodium.
}
\examples{
users <- data.frame(
  user = c("admin", "jan"),
  password = c("secret", "geheim")
)

set_app_users(
  users = users,
  hash_passwords = FALSE,
  renviron = ""
)
}
