% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{set_dropbox_token}
\alias{set_dropbox_token}
\title{Prepare Dropbox token for environment variable}
\usage{
set_dropbox_token(token_file = NULL, renviron = ".Renviron")
}
\arguments{
\item{token_file}{Path to token.rds. If not provided, `rdrop2::drop_auth()`
is called to generate a valid rds file for the current user.}

\item{renviron}{Path to .Renviron. If empty string is provided than only the
token is returned (no file modified).}
}
\value{
Character scalar.
}
\description{
Reads a token.rds, validates it and converts it to a
base64 string suitable for DROPBOX_TOKEN.
}
\examples{
\dontrun{
  x <- set_dropbox_token()
}
}
