% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{validate_user}
\alias{validate_user}
\title{Validate credentials}
\usage{
validate_user(user, password)
}
\arguments{
\item{user}{Username}

\item{password}{Password}
}
\value{
logical
}
\description{
Validate credentials
}
\examples{
\dontrun{
  validate_user("admin", "secret")
}
}
